MBS-20 Continue refactor
